requirements-test.txt
# Production requirements
-r requirements.txt

# Development/testing dependencies
pytest>=6.0.0
pytest-flask>=1.0.0